DIABETESPREDICTOR/requirements.txt
joblib==1.3.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
reportlab==4.1.0
streamlit==1.32.1
streamlit_option_menu==0.3.12
seaborn==0.13.2
scikit-learn==1.4.1.post1
plotly==5.20.0
imbalanced-learn==0.12.2